Bugfix and Feature Additions
Added auto fetch of Direct Message History for the logged in user, as well as saving it to the localStorage assigned to its own id. (will update to only save the user and not include the messages)

Also added auto navigate when a user already has a session, and restriction to dashboard when there isn't.

Fixed Signup Error Handling message
Updated api.ts